pyproject.toml
[project]
name = "forgecode"
version = "0.1.0"
description = "A Python library for in code code generation"
authors = [{ name = "Filip Mihajlović", email = "filipmihajlovic@icloud.com"}]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8.1,<4.0"

keywords = ["python", "library", "development", "forgecode"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[project.urls]
Homepage = "https://github.com/Cofyka/forgecode"
Repository = "https://github.com/Cofyka/forgecode"
Documentation = "https://Cofyka.github.io/forgecode"

# [project.dependencies] 

[tool.poetry]
packages = [{ include = "forgecode", from = "src" }]

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
pytest = "^7.2.0"
pre-commit = "^3.0.0"
python-dotenv = "^1.0.1"


[tool.poetry.dependencies]
openai = "^1.65.2"
jsonschema = "^4.23.0"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
